- parsing economy
- updating stock and price
- buy&sell screen (unfinished)
- improved city pathfinding (not yet used)
- research unlocks, building access and completion
- all city tabs at least partially functional

+++ framework/render/gles30_v2/ogles_3_0_renderer_v2.cpp
@@ -341,7 +341,7 @@ class Spritesheet
 		}
 		// Required a new page
 		LogInfo("Creating spritesheet page %d", (int)pages.size());
-		auto page = mksp<SpritesheetPage>(pages.size(), page_size, node_count);
+		auto page = mksp<SpritesheetPage>((int)pages.size(), page_size, node_count);
 		auto ret = page->addEntry(entry);
 		if (!ret)
 		{
